requirements.txt
alembic==1.14.0
amqp==5.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
greenlet==3.1.1
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.7
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
pymysql==1.1.1
cryptography
pytest
